requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
passlib==1.7.4
python-multipart==0.0.6
jinja2==3.1.2
email-validator==2.1.0.post1
python-dotenv==1.0.0
alembic==1.13.0
python-jose[cryptography]==3.3.0